fix(web): prevent login hydration mismatch
Move the login route shell back to a server component so PWA auth redirects render the same page tree on the server and client.
